+++ README.md
@@ -2,6 +2,9 @@
 
 ##Modification History
 
+###4 - Improve architecture using MVC Pattern
+- Create interface to respect SOLID principles
+
 ###3 - Migrate Gradle from Groovy to Kotlin Script
 - Adapt syntax to look like Kts
 - Upgrade Gradle to 7.0.2
